tests: moved tests into integration namespace
