+++ Onspring.API.SDK.Tests/Tests/Integration/OnspringClientReportsTests.cs
@@ -6,7 +6,7 @@ using Onspring.API.SDK.Tests.Infrastructure.Http;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 
-namespace Onspring.API.SDK.Tests.Tests
+namespace Onspring.API.SDK.Tests.Tests.Integration
 {
     [TestClass, ExcludeFromCodeCoverage]
     public class OnspringClientReportsTests
